moved useEffect for connection into active game room
